src/app/mail_menu/requirements.txt
flask
python-dotenv
Flask-SQLAlchemy
psycopg2-binary
PyJWT